refactor: SIMT-32 파일 이름 변경
