+++ feature/feature-auth/src/main/java/com/comit/feature_auth/navigation/AuthNavigation.kt
@@ -6,8 +6,8 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
 import com.comit.feature_auth.screen.find.AuthFindScreen
-import com.comit.feature_auth.screen.signin.SignInScreen
-import com.comit.feature_auth.screen.signup.SignUpScreen
+import com.comit.feature_auth.screen.signIn.SignInScreen
+import com.comit.feature_auth.screen.signUp.SignUpScreen
 import com.comit.navigator.SimTongRoute
 import com.comit.navigator.SimTongScreen
 import com.google.accompanist.pager.ExperimentalPagerApi
