+++ feature/feature-auth/src/main/java/com/comit/feature_auth/screen/signIn/SignInScreen.kt
@@ -1,7 +1,6 @@
-package com.comit.feature_auth.screen.signin
+package com.comit.feature_auth.screen.signIn
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
